internal: Move normalize files
